% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(data_path = NULL, meta_path = NULL, annotation_path = NULL)
}
\arguments{
\item{data_path}{character, path to folder containing data.}

\item{meta_path}{character, path to the data meta file.}

\item{annotation_path}{character, path to the annotation file.}
}
\description{
Run the Shiny Application
}
